requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.1
pillow>=10.0.0
opencv-python>=4.9.0.80
azure-cognitiveservices-vision-computervision>=0.9.0